Fix double to in conversion for omnetpp 551

+++ src/sdn4core/switch/avb/services/avb/SRP/OF_SRProtocol.cc
@@ -83,7 +83,7 @@ void OF_SRProtocol::handleMessage(cMessage *msg)
                 unsigned long utilizedBandwidth = srpTable->getBandwidthForModule(port);
                 //Add Higher Priority Bandwidth
                 utilizedBandwidth += static_cast<unsigned long>(port->getSubmodule("phy")->getSubmodule("shaper")->par(
-                        "AVBHigherPriorityBandwidth").longValue());
+                        "AVBHigherPriorityBandwidth"));
                 unsigned long requiredBandwidth = srpTable->getBandwidthForStream(listenerReady->getStreamID(),
                         listenerReady->getVlan_identifier());
